==> R_scripts/02_rate_generation_all_sets.Rmd <==
---
title: "Generating rates of change and CIs for all SETs"
author: "Kim Cressman"
date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: yes
  html_document:
    toc: yes
    toc_float: yes
always_allow_html: yes
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message = FALSE, warning = FALSE}

library(dplyr)
library(tidyr)
library(purrr)
library(readxl)
library(stringr)
library(lubridate)
library(here)
library(janitor)
library(nlme)
library(ggplot2)
library(forcats)
library(flextable)
```

Read in combined data file.  

```{r}
in_path <- here::here("data", "03_combined_processed", "all_sites.csv")
dat <- read.csv(in_path, stringsAsFactors = FALSE) %>% 
    mutate(date = lubridate::ymd(paste(year, month, day, sep = "-")),
           date2 = lubridate::decimal_date(date))
```



***

Read in metadata.  

```{r}
in_path <- here::here("data", "03_combined_processed", "all_sites_metadata.csv")
mdat <- read.csv(in_path, stringsAsFactors = FALSE)
```

Make sure all SETs we have in the `dat` data frame are present in the metadata, and that they have lat/long coordinates. Get rid of anything in the metadata that's not in the data (some reserves have information on SETs that aren't up and running yet).  

```{r}
# first pull out set_id from both data frames
data_setid <- unique(as.character(dat$set_id))
metadata_setid <- unique(mdat$set_id)
# find set_ids that are in the data, but not in the metadata
dat_not_m <- setdiff(data_setid, metadata_setid)
# find set_ids that are in the metadata, but not in the data
m_not_dat <- setdiff(metadata_setid, data_setid)

# print a message
if (length(dat_not_m) > 0) {
    toprint <- paste(dat_not_m, collapse = ", ")
    warning(paste0("The following SET IDs exist in your data, but not in your metadata: ", toprint))
}
```

From original SETr national workup, and still true on 6/9/2025:  

+  APA Little St. Marks was only measured once, in 2016. Will leave it in data file but remove it from data frame for this analysis.  
 


```{r}
if (length(m_not_dat) > 0) {
    toprint <- paste(m_not_dat, collapse = ", ")
    warning(paste0("The following SET IDs exist in your metadata, but not in your data, and will be removed within the current analysis: ", toprint))
}
if (length(dat_not_m) + length(m_not_dat) == 0) {
    print("SET IDs match in your data and metadata files.")
}
```


Weed out the APA station from the data; remove stations in metadata that aren't in data.  

```{r}
to_remove_dat <- c("Little St. Marks")
dat <- filter(dat, !(set_id %in% to_remove_dat))
mdat <- filter(mdat, !(set_id %in% m_not_dat))
```

Check for any stations in the metadata that don't have lat/long coordinates.  

```{r}
sum(is.na(mdat$lat))
mdat %>% 
  filter(is.na(lat)) %>% 
  select(reserve, set_id)
```

All fine as of 6/9/2025.

***

# Generate time series for later graphics  

Average all 36 within a date in order to make a small data frame.

```{r}
set_avgd_readings <- dat |> 
    group_by(reserve, set_id, year, month, day) |> 
    summarize(height_mean = mean(pin_height, na.rm = TRUE),
           height_sd = sd(pin_height, na.rm = TRUE)) |> 
    ungroup() |> 
    filter(!is.na(height_mean)) |> 
    arrange(reserve, set_id, year, month, day) |> 
    group_by(reserve, set_id) |> 
    mutate(height_cumu = height_mean - height_mean[1],
           date = lubridate::ymd(paste(year, month, day)),
           across(c(height_mean, height_sd, height_cumu), function(x) round(x, 2))) |> 
    ungroup() 
```


# Filter by time and number of samples  

Filter to only analyze sites with at least 5 measurements over at least 4.5 years.  

```{r}
to_keep <- dat %>% 
       group_by(reserve, set_id, arm_position, pin_number) %>% 
        summarize(count = n(),
                  daterange = max(date2) - min(date2)) %>% 
        filter(count >= 5,
               daterange >= 4.5)


to_analyze <- dat %>% 
        filter(set_id %in% unique(to_keep$set_id))
```

Working with how many SETs?  

```{r}
length(unique(to_analyze$set_id))
```

At how many reserves?  

```{r}
length(unique(to_analyze$reserve))
```

# Generate rates  

Run random-intercept LMMs as in the main project: fixed effect = date; random effects = pin nested within arm.  

```{r}
models2 <- to_analyze %>%
    group_by(reserve, set_id) %>%
    do(mod = lme(pin_height ~ date, data = ., random = ~1|arm_position/pin_number, na.action = na.omit))
```

Pull out information from the models.  

```{r}
lmm_out <- models2 %>% 
  mutate(rate = intervals(mod, which = "fixed")$fixed["date", 2] * 365.25,
         CI_low = intervals(mod, which = "fixed")$fixed["date", 1] * 365.25,
         CI_high = intervals(mod, which = "fixed")$fixed["date", 3] * 365.25) %>% 
  select(-mod)
```

```{r}
lmm_out %>% 
  flextable() %>% 
  theme_booktabs() %>% 
  autofit()
```

***
***

# Join rates to metadata  

Also pull out dates.  

```{r}
all_dates <- dat |> 
    summarize(.by = c(reserve, set_id),
              start_date = min(date),
              end_date = max(date),
              ts_length = round(max(date2) - min(date2), 1))
```

Join them to metadata.  

```{r}
mdat <- full_join(mdat, all_dates)
```


Need to do a full join; there are some sites in the data that aren't in metadata; and some sites in metadata that didn't have enough data to be analyzed but we can put them on a map with some icon.  

```{r}
all_out <- full_join(mdat, lmm_out, by = c("reserve", "set_id"))
```


Get rid of any sites that are missing *both* a rate and a location.  (commented out)

```{r}
# # find 'em
# ditch <- all_out %>% 
#   filter(is.na(lat),
#          is.na(rate)) %>% 
#   select(reserve, set_id)
# 
# # ditch 'em
# all_out <- all_out %>% 
#   filter(!(set_id %in% ditch$set_id))
```


Read in SLR rates and join them to the data frame so we can make comparisons. Also read in reserve information.   

```{r}
slr_path <- here::here("data", "01b_reserve_metadata", "slr_rates.csv")
slr_rates <- read.csv(slr_path, stringsAsFactors = FALSE) %>% 
  clean_names() %>% 
  remove_empty(which = c("rows", "cols"))

to_join <- slr_rates %>% 
  mutate(slr_CI_low = slr_rate_mm_yr - x_95_ci,
         slr_CI_high = slr_rate_mm_yr + x_95_ci,
         yr19_CI_low = slr_19yr - ci95_19yr,
         yr19_CI_high = slr_19yr + ci95_19yr) %>% 
  select(reserve, 
         slr_rate = slr_rate_mm_yr,
         slr_CI_low,
         slr_CI_high,
         slr_19yr,
         yr19_CI_low,
         yr19_CI_high,
         slr_start = data_start,
         yr19_start,
         slr_end = data_end)

all_to_compare <- left_join(all_out, to_join, by = "reserve") 
```

Weed out some columns; generate comparison columns for both 0 and slr.  

```{r}
set_details <- all_to_compare %>% 
    mutate(
        # paste together veg columns, omitting NAs
        dominant_vegetation = pmap_chr(
            list(co_dominant_species1, co_dominant_species2, co_dominant_species3, co_dominant_species4),
            ~ paste(na.omit(c(...)), collapse = ", ")),
        # replace empty strings with "not provided"
        dominant_vegetation = case_when(dominant_vegetation == "" ~ "Not provided",
                                        .default = dominant_vegetation),
        # deal with "deep rod set" consistency
        set_type = case_when(set_type == "Deep ROD SET" ~ "Deep Rod SET",
                             .default = set_type),
        # deal with lack of user friendly set names
        user_friendly_set_name = case_when(is.na(user_friendly_set_name) ~ set_id,
                                           .default = user_friendly_set_name),
        # deal with lack of salinity
        general_salinity = case_when(is.na(general_salinity) ~ "Not provided",
                                     .default = general_salinity),
        # round SLR rates and CIs
        across(c(rate, CI_low, CI_high,
         slr_rate, slr_CI_low, slr_CI_high,
         slr_19yr, yr19_CI_low, yr19_CI_high), function(x) round(x, 3))
    ) |> 
  select(reserve, 
         set_id, 
         user_friendly_set_name, 
         lat, long, 
         set_type,
         dominant_vegetation,
         general_salinity,
         start_date, end_date, ts_length,
         rate, CI_low, CI_high,
         slr_rate, slr_CI_low, slr_CI_high,
         slr_19yr, yr19_CI_low, yr19_CI_high) %>% 
  mutate(dir_0 = case_when(CI_high < 0 ~ "dec_sig",
                           CI_low > 0  ~ "inc_sig",
                           rate < 0 ~ "dec_nonsig",
                           rate > 0 ~ "inc_nonsig",
                           TRUE ~ "not_enough_info"),
         dir_slr = case_when(CI_high < slr_CI_low ~ "dec_sig",
                             CI_low > slr_CI_high  ~ "inc_sig",
                             rate < slr_rate ~ "dec_nonsig",
                             rate > slr_rate ~ "inc_nonsig",
                             TRUE ~ "not_enough_info"),
         dir_19yr = case_when(CI_high < yr19_CI_low ~ "dec_sig",
                              CI_low > yr19_CI_high  ~ "inc_sig",
                              rate < slr_19yr ~ "dec_nonsig",
                              rate > slr_19yr ~ "inc_nonsig",
                              TRUE ~ "not_enough_info"),
         set_to_slr = round(rate / slr_rate, 3),
         set_to_19yr = round(rate / slr_19yr, 3)) 
```

```{r}
reserve_details <- read.csv(here::here("data",
                                       "NERR Websites.csv"))

reserve_sets <- set_details |> 
    select(reserve, set_id, dir_0, dir_slr, dir_19yr) |> 
    pivot_longer(cols = c(dir_0, dir_slr, dir_19yr),
                 names_to = "compared_to",
                 names_prefix = "dir_",
                 values_to = "outcome") |> 
    summarize(.by = c(reserve, compared_to, outcome),
              n = n()
              ) |> 
    complete(reserve, compared_to, outcome,
             fill = list(n = 0))
```

```{r}
# set up the nicer names for comparisons
set_details <- set_details |> 
    mutate(dir_19yr = factor(dir_19yr, 
                             levels = c("dec_sig", "dec_nonsig", "not_enough_info", "inc_nonsig", "inc_sig"),
                             labels = c("No, more confident", "No, less confident", "Not calculated",
                                        "Yes, less confident", "Yes, more confident")),
           dir_slr = factor(dir_slr, 
                            levels = c("dec_sig", "dec_nonsig", "not_enough_info", "inc_nonsig", "inc_sig"),
                            labels = c("No, more confident", "No, less confident", "Not calculated",
                                       "Yes, less confident", "Yes, more confident")))



# make a list of reserve summary details
# make the reserve-level list
reserves <- unique(set_details$reserve)
reserves_sets_list <- list()

# for each reserve, make a list of different SET characteristics to report
for(i in seq_along(reserves)){
    
    res_set_details <- set_details |> 
        filter(reserve == reserves[i])
    
    # general details of SET landscapes and settings
    set_total <- res_set_details |> 
        summarize(.by = reserve,
                  n = n())
    
    set_installations <- res_set_details |>
        mutate(start_year = lubridate::year(start_date)) |> 
        summarize(.by = c(reserve, start_year),
                  n = n()) |> 
        arrange(reserve, start_year)
    set_types <- res_set_details |> 
        summarize(.by = c(reserve, set_type),
                  n = n())
    set_vegs <- res_set_details |> 
        summarize(.by = c(reserve, dominant_vegetation),
                  n = n())
    
    # SLR rates for comparison
    set_slr_rates <- res_set_details |> 
        select(reserve, slr_rate, slr_CI_low, slr_CI_high,
               slr_19yr, yr19_CI_low, yr19_CI_high) |> 
        mutate(across(c(slr_rate, slr_CI_low, slr_CI_high,
               slr_19yr, yr19_CI_low, yr19_CI_high),
               function(x) round(x, 2))) |> 
        summarize(.by = c(reserve, slr_rate, slr_CI_low, slr_CI_high,
               slr_19yr, yr19_CI_low, yr19_CI_high),  
               n = n()) |> 
        mutate(longterm = glue::glue("{slr_rate} mm/yr ({slr_CI_low} - {slr_CI_high})"),
               yr19 = glue::glue("{slr_19yr} mm/yr ({yr19_CI_low} - {yr19_CI_high})")) |> 
    select(reserve, n, slr_rate, slr_19yr, longterm, yr19) |> 
    pivot_longer(c(longterm, yr19),
                 names_to = "comp",
                 values_to = "Rate") |> 
    mutate(comp = case_when(comp == "longterm" ~ "Long-term",
                            comp == "yr19" ~ "Near-term"),
           Water_Rate = case_when(comp == "Long-term" ~ slr_rate,
                                  comp == "Near-term" ~ slr_19yr)) |> 
        rename(Water_Comparison = comp) |> 
        select(reserve, Water_Comparison, Water_Rate, Rate, n)
    
    # summary of comparisons
    slr_comps <- res_set_details |> 
        summarize(.by = c(dir_slr, slr_rate),
                  n = n()) |> 
        rename(Result = dir_slr,
               Water_Rate = slr_rate) |> 
        mutate(Water_Comparison = "Long-term")
    
    yr19_comps <- res_set_details |> 
        summarize(.by = c(dir_19yr, slr_19yr),
                  n = n()) |> 
        rename(Result = dir_19yr,
               Water_Rate = slr_19yr) |> 
        mutate(Water_Comparison = "Near-term")
    
    all_comps <- bind_rows(slr_comps, yr19_comps)
    
    comps_wide_withRates <- all_comps |> 
        arrange(desc(Result)) |> 
        pivot_wider(names_from = Result,
                    values_from = n,
                    values_fill = 0) |> 
        as.data.frame() |> 
        left_join(set_slr_rates) |> 
        relocate(Water_Comparison, Rate) |> 
        select(-reserve, -Water_Rate) |> 
        rename(Total = n)
        
    
    
    # salinity
    set_salinities <- res_set_details |> 
        summarize(.by = c(reserve, general_salinity),
                  n = n()) 
    
    
    # combine
    res_set_list <- list(
        total = set_total,
        installations = set_installations,
        types = set_types,
        vegs = set_vegs,
        water_comp_rates = set_slr_rates,
        water_comp_summaries = comps_wide_withRates,
        salinities = set_salinities
    )
    
    reserves_sets_list[[i]] <- res_set_list
    names(reserves_sets_list)[i] <- reserves[i]
}

```


Write out.  

```{r}
save(reserve_sets, reserve_details, set_details,
     reserves_sets_list, set_avgd_readings,
     file = here::here("data",
                       "04_rates",
                "SET_rates_and_details.RData"),
     compress = "xz")
```

`SET_rates_and_details.RData` is what I'll want to use in the app.
